Fix Paginator returning 0 total pages.
If the Paginator class was passed 0 for the total number of
items, it would return 0 for the total number of pages. This
could cause an infinite redirect problem if you choose to redirect
the user if they request a page number larger than the total
number of pages. This fix ensures that the total number of pages
is 1 if the total number of items is 0.