Fix back link being set to favicon.ico.
'back' was being set in the session each time the router handles
a get request, including those from ajax. This meant the last
get request processed was set as the 'back' instead of the main
url. This fix ensures the 'back' link is only set when the
requested url matches a registered route.